refactor(recovery): split terminal outcome helpers
